Rename RogueClient as GatewayClient
