+++ server/config/lib/client.js
