+++ server/lib/endpoint-signups.js
